+++ LAB1/requirements.md
@@ -1,0 +1,5 @@
+I6B1S1
+Damian TOMASIK
+
+Aby uruchomić program należy używając Visual Studio 2017 lub nowzego pobrać pakier OR-TOOLS przy pomocy NuGet.
+Następnie należy uruchomić program.
