+++ LAB2/65459/OtoDomScraper/bin/Debug/netcoreapp2.1/OtoDomScraper.runtimeconfig.dev.json
@@ -1,0 +1,10 @@
+{
+  "runtimeOptions": {
+    "additionalProbingPaths": [
+      "C:\\Users\\Damian\\.dotnet\\store\\|arch|\\|tfm|",
+      "C:\\Users\\Damian\\.nuget\\packages",
+      "C:\\Microsoft\\Xamarin\\NuGet",
+      "C:\\Program Files\\dotnet\\sdk\\NuGetFallbackFolder"
+    ]
+  }
+}
