--- LAB2/OtoDomScraper.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29009.5
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OtoDomScraper", "OtoDomScraper\OtoDomScraper.csproj", "{B91C5045-DE52-40F6-B593-E86ACB8EC9C8}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B91C5045-DE52-40F6-B593-E86ACB8EC9C8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B91C5045-DE52-40F6-B593-E86ACB8EC9C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B91C5045-DE52-40F6-B593-E86ACB8EC9C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B91C5045-DE52-40F6-B593-E86ACB8EC9C8}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1C4F5EE5-E30C-4D3E-9004-B2C990684F51}
-	EndGlobalSection
-EndGlobal
