--- LAB2/OtoDomScraper/bin/Debug/netcoreapp2.1/OtoDomScraper.runtimeconfig.dev.json
@@ -1,10 +1,0 @@
-{
-  "runtimeOptions": {
-    "additionalProbingPaths": [
-      "C:\\Users\\Damian\\.dotnet\\store\\|arch|\\|tfm|",
-      "C:\\Users\\Damian\\.nuget\\packages",
-      "C:\\Microsoft\\Xamarin\\NuGet",
-      "C:\\Program Files\\dotnet\\sdk\\NuGetFallbackFolder"
-    ]
-  }
-}
