--- LAB2/OtoDomScraper/bin/Debug/netcoreapp2.1/OtoDomScraper.runtimeconfig.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "tfm": "netcoreapp2.1",
-    "framework": {
-      "name": "Microsoft.NETCore.App",
-      "version": "2.1.0"
-    }
-  }
-}
