+++ LAB2/README.md
@@ -1,0 +1,3 @@
+Aby uruchomić scraper, należy otworzyć projekt w Visual Studio i przy pomocy NuGet pobrać:
+AngleSharp oraz HtmlAgilityPack
+Program skompilować i uruchomić przy pomocy środowiska
