--- requirements.md
@@ -1,5 +1,0 @@
-I6B1S1
-Damian TOMASIK
-
-Aby uruchomić program należy używając Visual Studio 2017 lub nowzego pobrać pakier OR-TOOLS przy pomocy NuGet.
-Następnie należy uruchomić program.
